+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.0
 - Unit Tests!
 - Requires ZEsarUX 7.3.
+- Supports new breakpoint condition format of ZEsarUX v7.3.
 
 ## 0.8.1
 - Memory pages are shown in VARIABLES section.